# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankara Eye Hospital Vision 2020 Conference 2026 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance/scan:
    post:
      operationId: scanAttendance
      tags: [attendance]
      summary: Scan QR for attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceScanInput"
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceScanResult"
  /attendance/logs:
    get:
      operationId: listAttendanceLogs
      tags: [attendance]
      summary: List attendance logs
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceLog"

  # DASHBOARD
...
components:
  schemas:
...
    Participant:
      type: object
      required: [id, registrationNumber, name, email, mobile, institution, createdAt, isPaid, isActive, isSponsored, delegateType]
      properties:
        id:
          type: integer
        registrationNumber:
          type: string
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        institution:
          type: string
        createdAt:
          type: string
          format: date-time
        hasPassword:
          type: boolean
        isPaid:
          type: boolean
        utrNumber:
          type: string
          nullable: true
        isActive:
          type: boolean
        isSponsored:
          type: boolean
        sponsorType:
          type: string
          nullable: true
        delegateType:
          type: string
        roles:
          type: array
          items:
            type: string
...
    AttendanceScanInput:
      type: object
      required: [registrationNumber, day]
      properties:
        registrationNumber:
          type: string
        day:
          type: string
    AttendanceScanResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        status:
          type: string
          enum: [already_marked, already_collected, not_linked, not_onboarded, not_found]
        goodiesCollected:
          type: boolean
        participant:
          $ref: "#/components/schemas/Participant"
          nullable: true
    AttendanceLog:
      type: object
      required: [id, participantId, scannedAt]
      properties:
        id:
          type: integer
        participantId:
          type: integer
        participantName:
          type: string
        registrationNumber:
          type: string
        institution:
          type: string
        scannedAt:
          type: string
          format: date-time

    # DASHBOARD SCHEMAS
...